Added uuid suffix to ensure pageId is unique when there is a single page in the front or body
This is in preparation of following commits.
There will be an EditionSource selector and in case of just page_front/page1 as id, there are some checks in the code that prevent updates,
for example, at the time of this commit, in TextPanelComponent at line 80 there is a distinctUntilChanged that checks for pageId equality.
If the EditionSource changes and the new one also has just one page in its front or body, the pageId would be page_front/page1 once again and the pipeline would stop at that distinctUntilChanged.
Adding a uiid as suffix solve this easily, also I did not find any other places where such hardcoded ids where referenced, so it shouldn't break anything.
In the component I mentioned above, I avoided changing the ids check to JSON.stringify the page content, as in other places,
because that pipeline runs multiple times when scrolling.